string.h prototype matches declaration in string.c

+++ include/string.h
@@ -1,7 +1,7 @@
 #ifndef _STRING_H
 #define _STRING_H
 
-unsigned int strlen(const char* str);
+size_t strlen(const char* str);
 int strequ(const char* s1, const char* s2);
 
 #endif /* _STRING_H */
